Track management, fractional seconds, finding same person on different cams
Problem with fractional seconds values solved!
Trajectory management for all the tracks
Coding started to identify a particular person from different cameras using spatial (x,y) and temporal (time) variables.
Note: Currently the code will not run as it is. There is a problem in intercamera.R probably at line 44.

analyzetracks.R is supposed to be a sub-wrapper script.
main.R is supposed to be a single wrapper script for all processing in R.