Added code to define stops for visitors in a track
Script added to find stops in a visitor's track according to a parameter which can be adjusted manually or made dynamic by looking at statistics of each track (using standard deviations for the distance). We use a spatial (distance) window of 0.5m and a temporal window of 1.5 seconds (3 consecutive observations)

We also visualize these stops on the xy plot in the global coordinate space using transparent rectangles.

Note: analyzetracks.R and its associated files are not complete and therefore will throw errors currently.